+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.7.2
 
-  - build templated path
+  - build templated path (see bootstrap/standalone/project.xml)
   - centralize toolchain import modules
 
 ## 2.7.1